+++ src/core/abc/session.abc.ts
@@ -43,6 +43,7 @@ export function getBrowserExecutablePath() {
   }
   return CHROMIUM_PATH;
 }
+
 export function ensureSuffix(phone) {
   const suffix = '@c.us';
   if (phone.includes('@')) {
@@ -168,9 +169,34 @@ export abstract class WhatsappSession {
 
   abstract stopTyping(chat: ChatRequest);
 
-  abstract getMessages(query: GetMessageQuery);
+  getMessages(query: GetMessageQuery) {
+    throw new NotImplementedByEngineError();
+  }
 
   abstract setReaction(request: MessageReactionRequest);
+
+  /**
+   * Chats methods
+   */
+  public getChats() {
+    throw new NotImplementedByEngineError();
+  }
+
+  public deleteChat(chatId) {
+    throw new NotImplementedByEngineError();
+  }
+
+  public getChatMessages(
+    chatId: string,
+    limit: number,
+    downloadMedia: boolean,
+  ) {
+    throw new NotImplementedByEngineError();
+  }
+
+  public clearMessages(chatId) {
+    throw new NotImplementedByEngineError();
+  }
 
   /**
    * Contacts methods